refactor(data): centralize phase-1 DB access into data crate
Move direct diesel queries out of the server crate and into palpo-data,
keeping business logic (regex matching, UIAA flow, signing-key merge,
retry backoff) in the server while the data layer owns all SQL.

New data modules:
- data::appservice    - appservice_registrations CRUD + DbRegistration model
- data::user::uiaa    - user_uiaa_datas session/request storage
- data::room::transaction_id - event_idempotents txn-id lookups

Extended data modules:
- data::misc          - server_signing_keys read/upsert helpers
- data::sending       - persist_retry_state for outgoing_requests

Server modules (appservice, uiaa, transaction_id, server_key, sending::guard)
now delegate to the data layer and no longer reference connect()/schema/diesel.
DbRegistration is re-exported from crate::appservice for source compatibility.
