+++ crates/server/src/appservice.rs
@@ -1,19 +1,13 @@
 use std::collections::BTreeMap;
-use std::fmt;
 use std::time::Duration;
 
-use diesel::prelude::*;
-use diesel_async::RunQueryDsl;
 use regex::RegexSet;
-use serde::{Deserialize, Serialize};
 use subtle::ConstantTimeEq;
 
 use crate::core::appservice::{Namespace, Registration};
 use crate::core::identifiers::*;
-use crate::core::serde::JsonValue;
-use crate::data::connect;
-use crate::data::schema::*;
-use crate::{AppError, AppResult, sending};
+pub use crate::data::appservice::DbRegistration;
+use crate::{AppError, AppResult, data, sending};
 
 /// Compiled regular expressions for a namespace.
 #[derive(Clone, Debug)]
@@ -133,151 +127,10 @@ impl TryFrom<DbRegistration> for RegistrationInfo {
     }
 }
 
-#[derive(Identifiable, Queryable, Insertable, Serialize, Deserialize, Clone)]
-#[diesel(table_name = appservice_registrations)]
-pub struct DbRegistration {
-    /// A unique, user - defined ID of the application service which will never change.
-    pub id: String,
-
-    /// The URL for the application service.
-    ///
-    /// Optionally set to `null` if no traffic is required.
-    pub url: Option<String>,
-
-    /// A unique token for application services to use to authenticate requests to HomeServers.
-    pub as_token: String,
-
-    /// A unique token for HomeServers to use to authenticate requests to application services.
-    pub hs_token: String,
-
-    /// The localpart of the user associated with the application service.
-    pub sender_localpart: String,
-
-    /// A list of users, aliases and rooms namespaces that the application service controls.
-    pub namespaces: JsonValue,
-
-    /// Whether requests from masqueraded users are rate-limited.
-    ///
-    /// The sender is excluded.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub rate_limited: Option<bool>,
-
-    /// The external protocols which the application service provides (e.g. IRC).
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub protocols: Option<JsonValue>,
-
-    /// Whether the application service wants to receive ephemeral data.
-    ///
-    /// Defaults to `false`.
-    pub receive_ephemeral: bool,
-
-    /// Whether the application service wants to do device management, as part of MSC4190.
-    ///
-    /// Defaults to `false`
-    #[serde(default, rename = "io.element.msc4190")]
-    pub device_management: bool,
-
-    /// Whether this appservice is administratively disabled.
-    ///
-    /// Disabled appservices are loaded but not returned by `all()`, so they
-    /// neither receive events nor authenticate requests.
-    #[serde(default)]
-    pub disabled: bool,
-}
-
-// Custom Debug implementation to prevent leaking as_token and hs_token
-impl fmt::Debug for DbRegistration {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        f.debug_struct("DbRegistration")
-            .field("id", &self.id)
-            .field("url", &self.url)
-            .field("as_token", &"[REDACTED]")
-            .field("hs_token", &"[REDACTED]")
-            .field("sender_localpart", &self.sender_localpart)
-            .field("namespaces", &self.namespaces)
-            .field("rate_limited", &self.rate_limited)
-            .field("protocols", &self.protocols)
-            .field("receive_ephemeral", &self.receive_ephemeral)
-            .field("device_management", &self.device_management)
-            .field("disabled", &self.disabled)
-            .finish()
-    }
-}
-
-impl From<Registration> for DbRegistration {
-    fn from(value: Registration) -> Self {
-        let Registration {
-            id,
-            url,
-            as_token,
-            hs_token,
-            sender_localpart,
-            namespaces,
-            rate_limited,
-            protocols,
-            receive_ephemeral,
-            device_management,
-        } = value;
-        Self {
-            id,
-            url,
-            as_token,
-            hs_token,
-            sender_localpart,
-            namespaces: serde_json::to_value(namespaces).unwrap_or_default(),
-            rate_limited,
-            protocols: protocols
-                .map(|protocols| serde_json::to_value(protocols).unwrap_or_default()),
-            receive_ephemeral,
-            device_management,
-            disabled: false,
-        }
-    }
-}
-impl TryFrom<DbRegistration> for Registration {
-    type Error = serde_json::Error;
-
-    fn try_from(value: DbRegistration) -> Result<Self, Self::Error> {
-        let DbRegistration {
-            id,
-            url,
-            as_token,
-            hs_token,
-            sender_localpart,
-            namespaces,
-            rate_limited,
-            protocols,
-            receive_ephemeral,
-            device_management,
-            disabled: _,
-        } = value;
-        let protocols = if let Some(protocols) = protocols {
-            serde_json::from_value(protocols)?
-        } else {
-            None
-        };
-        Ok(Self {
-            id,
-            url,
-            as_token,
-            hs_token,
-            sender_localpart,
-            namespaces: serde_json::from_value(namespaces)?,
-            rate_limited,
-            protocols,
-            receive_ephemeral,
-            device_management,
-        })
-    }
-}
-
 /// Registers an appservice and returns the ID to the caller
 pub async fn register_appservice(registration: Registration) -> AppResult<String> {
     let db_registration: DbRegistration = registration.into();
-    diesel::insert_into(appservice_registrations::table)
-        .values(&db_registration)
-        .execute(&mut connect().await?)
-        .await?;
+    data::appservice::insert_registration(&db_registration).await?;
     Ok(db_registration.id)
 }
 
@@ -287,26 +140,18 @@ pub async fn register_appservice(registration: Registration) -> AppResult<String
 ///
 /// * `service_name` - the name you send to register the service previously
 pub async fn unregister_appservice(id: &str) -> AppResult<()> {
-    diesel::delete(appservice_registrations::table.find(id))
-        .execute(&mut connect().await?)
-        .await?;
+    data::appservice::delete_registration(id).await?;
     Ok(())
 }
 
 /// Set the `disabled` flag on an appservice. Returns true if a row was updated.
 pub async fn set_appservice_disabled(id: &str, disabled: bool) -> AppResult<bool> {
-    let affected = diesel::update(appservice_registrations::table.find(id))
-        .set(appservice_registrations::disabled.eq(disabled))
-        .execute(&mut connect().await?)
-        .await?;
-    Ok(affected > 0)
+    Ok(data::appservice::set_disabled(id, disabled).await?)
 }
 
 /// List all registrations in the database, including disabled ones.
 pub async fn list_all_registrations() -> AppResult<Vec<(DbRegistration, bool)>> {
-    let regs = appservice_registrations::table
-        .load::<DbRegistration>(&mut connect().await?)
-        .await?;
+    let regs = data::appservice::all_registrations().await?;
     Ok(regs
         .into_iter()
         .map(|r| {
@@ -317,12 +162,7 @@ pub async fn list_all_registrations() -> AppResult<Vec<(DbRegistration, bool)>> 
 }
 
 pub async fn get_registration(id: &str) -> AppResult<Option<Registration>> {
-    if let Some(registration) = appservice_registrations::table
-        .find(id)
-        .first::<DbRegistration>(&mut connect().await?)
-        .await
-        .optional()?
-    {
+    if let Some(registration) = data::appservice::find_registration(id).await? {
         Ok(Some(registration.try_into()?))
     } else {
         Ok(None)
@@ -375,10 +215,7 @@ pub async fn is_exclusive_room_id(room_id: &RoomId) -> AppResult<bool> {
 }
 
 pub async fn all() -> AppResult<BTreeMap<String, RegistrationInfo>> {
-    let registrations = appservice_registrations::table
-        .filter(appservice_registrations::disabled.eq(false))
-        .load::<DbRegistration>(&mut connect().await?)
-        .await?;
+    let registrations = data::appservice::enabled_registrations().await?;
     Ok(registrations
         .into_iter()
         .filter_map(|db_registration| {
